pyproject.toml
[project]
name = "series-cross-link-suggester"
version = "0.1.0"
description = "Find internal cross-link opportunities across blog series posts"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "local-first-common",
    "python-frontmatter>=1.1.0",
    "typer>=0.24.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]

[tool.uv.sources]
local-first-common = { git = "https://github.com/jamalhansen/local-first-common.git", rev = "main" }

[tool.pytest.ini_options]
pythonpath = ["src"]
